Implementation of an explicit random choice between all the solution with that got the best current score.
Before it was considered that since the solution are randomly generated, they are already in a random order so no need to consider the solution with same scores. Just take the first best one. Now with this patch it is not like this anymore.
I did it because with the quality measures taken, it's important to choose randomly between the best solution. Score_files same sol for free and for pred ?